V2

+++ dsworkflow/move_datafiles.py
@@ -29,8 +29,11 @@ def release_lock(lockpath):
     lockpath.unlink()
 
 def get_year(datestr):
-    return dt.datetime.strptime(datestr, '%y%m%d').strftime('%Y')
-
+    twodigyear = int(datestr[:2])
+    if twodigyear < 40:
+        return str(2000 + twodigyear)
+    return str(1900 + twodigyear)
+    
 def files_ready(dir):
     era5files = list(dir.glob(f"{PREFIX}*"))
     lockfiles = list(dir.glob(f"{MOVELOCKFILE}*"))
@@ -88,7 +91,15 @@ if __name__ == '__main__':
     else:
         with open(here / LOCKDIR / DONEFN, "r") as src:
             for fn in src:
-                print(f"Deleting {fn.rstrip()}")
-                shutil.rmtree((here.parent / 'WRF' / fn.rstrip()))
-
+                if (
+                    (here.parent / 'WRF' / fn.rstrip()).is_dir() and
+                    len(fn.rstrip()) == 6):
+                    print(f"Deleting {fn.rstrip()}")
+                    try:
+                        shutil.rmtree((here.parent / 'WRF' / fn.rstrip()))
+                    except OSError as err:
+                        print(f"There was an error trying to delete {fn.rstrip()}")
+                        print("Check deletion file and proceed manually.")
+                        print(err)
+                        break
     release_lock(lockpath)
